+++ CHANGELOG.markdown
@@ -2,6 +2,8 @@
 
 ## unreleased
 
+* Use railsy config paths `config/image_optim.yml` and `config/image_optim/#{RAILS_ENV}.yml` and cache to `tmp/cache/image_optim` by default [@toy](https://github.com/toy)
+
 ## v0.1.0 (2016-08-14)
 
 * Extracted from [image_optim gem](https://github.com/toy/image_optim) [toy/image_optim#127](https://github.com/toy/image_optim/issues/127) [@toy](https://github.com/toy)